refactor(redskilled): decompose dashboard adapters
